+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+#### [1.2.0] - 2020-10-09
+
+##### Added
+- Redirect to 404 page if post or category not found
+
+#### [1.1.1] - 2020-10-09
+
+##### Added
+- Removed moment.js - replaced with day.js
+
 #### [1.1.0] - 2020-08-20
 
 ##### Added